Removed the Hardcoded values for calling the URL and added a REGEX to handle the URL and cut it down to the specific code that is required.
Also updated the Brackets so that they would display a bit better but this still really needs some better changes.

Added more code to handle the tournement information.

Added checks for the Check In-State for the players, this still needs to be sorted so that it can be unchecked incase someone accidentally presses the wrong person for check in.