Made changes so that the text box will disable for the first player when the match is complete but this needs to be edited for the the second player as well.
The next thing to do is store a local copy of the matches
from the API call rather than calling it every time we use it because
it hangs too much. Need to stop relying on the API Calls so much and
start creating local variables for each tournement and only update it when
required. Will need to look in to the best way of achiving this.

I also need to try and clean up Bracket.razor becausethe auto format
doesn't work currently in VSCode and I know it's to do with the
blazor layout currently or something within the blazor HTML
as I had this previously with the honours project.

ALso added some additional bracket files that are currenty blank as
I want to move to using components to help clean up the code as the
razor files are starting to get a bit big and messy.