We can now modify the data on challonge and have it sync up with what is on there. This means it can now potentially add players from the Cehck-in screen if we are adding new players from there instead of on challonge.
We also now have access to google sheets, though this is still very wip and needs a lot of work. We can get all the data and sheet names, but because of the formatting differences between the sheets, we run in to the problem of the data being different across them all. I need to find a way of dealing with the potential of different formats as well.

Cleaned up the Checkin.razor code section to remoe some repeating code so that we just call the one method when adding a player to the list. This gets cleared off each time is called again to populate the list properly.

Bracket.razor has been modified so that the Confirm and edit buttons appear depending on what the current state of the match is. Confirm will confirm the entered results and update the Challonge and Edit will re-open the match to allow for the match to be modified again if the results where entered incorrectly.

Updated Appsettings.cs for further settings I want to add to the doc, need to set up a sheet selection so it can be saved.

Players.cs was also updated to help store the infromation that also comes in from the google sheets, this should allow for the information to be tied together when we want to update the leaderboard.