Optimised the google sheets stuff a bit more, it now loads when the application loads/page refreshes and will keep them stored allowing for quick switching when looking at the leaderboard page. I think this will still need some deeper optimisation like saving the Sheet data locally as well and checking on launch/refresh if we need to update the data.
Started adding the components in to the checkin page to add URLs and select previously selected URLs. I need to cache these better too again, adding these to the local data will help as well rather than having to call the API.

NExt stage will be to finish creating a new tournament through the applciation and get it to create a new sheet based on the new tournament.