Google login in is now working and keeps the access token active for 30 days. I've made it so that when the Access Token is Generated it gets saved to AccessToken.json in the application files, but is encrypted and is decrypted when trying to access the API at the start of the application to ensure the connection ad we set the token for the HttpContext. Next thing is to setup the Refresh token so that when the 30 Days is up we can refresh the connection.
Updated the Goolge sheet format to follow the new format that has been added on the Unoffical UK leaderboard, unfortunately google hasn't added Tables to the API, so I will need to sort the order through the app itself which currentl I'm struggling to do with the additional columns but should be able to access it soon.

Made it so that when the application is launched for the first time/when app.user.settings.json is missing it will only show the Settings screen and when you save on the settings screen it will redriect you to the checkin page. I want to update this as it feels a bit clunky.

Added a loading wheel thingy to the Check in page to indicate when there are things happening. Usually appears at the start of the app when trying to cache all the spreadsheets.